Feature/corona

+++ weather-data-api/src/main/java/com/weather/weatherdataapi/service/CoronaService.java
@@ -1,5 +1,7 @@
 package com.weather.weatherdataapi.service;
 
+import com.weather.weatherdataapi.exception.AlreadyExistsLatestDataException;
+import com.weather.weatherdataapi.exception.FailedFetchException;
 import com.weather.weatherdataapi.model.dto.responsedto.ScoreResultResponseDto;
 import com.weather.weatherdataapi.model.dto.responsedto.WeatherDataResponseDto;
 import com.weather.weatherdataapi.model.entity.BigRegion;
@@ -55,16 +57,16 @@ public class CoronaService {
     }
 
     public void setInfoAndScore(BigRegion currentBigRegion, ScoreResultResponseDto scoreResultResponseDto, WeatherDataResponseDto weatherDataResponseDto) {
-        CoronaInfo coronaInfo = getLatestInfoByBigRegion(currentBigRegion);
+        CoronaInfo coronaInfo = getInfoByBigRegion(currentBigRegion);
 
         weatherDataResponseDto.setCoronaCurrentBigRegionNewCaseCount(coronaInfo.getNewLocalCaseCount() + coronaInfo.getNewForeignCaseCount());
 
         weatherDataResponseDto.setCoronaAllNewCaseCount(getAllNewCaseCount());
 
-        convertInfoToScore(coronaInfo, scoreResultResponseDto);
+        convertInfoToScore(scoreResultResponseDto);
     }
 
-    public CoronaInfo getLatestInfoByBigRegion(BigRegion bigRegion) {
+    public CoronaInfo getInfoByBigRegion(BigRegion bigRegion) {
         CoronaInfo coronaInfo;
 
         // 캐시 데이터가 있다면 캐시 데이터를 먼저 사용합니다.
@@ -83,12 +85,15 @@ public class CoronaService {
     @Transactional
     public void fetchAndStoreCoronaInfoUsingOpenApi() throws Exception {
         if (checkAlreadyHasLatestData() == true)
-            return;
+            throw new AlreadyExistsLatestDataException();
+
+        ICoronaInfo info = govCoronaOpenApi.getInfo();
+
+        if (info.getItemList().isEmpty())
+            throw new FailedFetchException("가져온 아이템 리스트가 비어있습니다.");
 
         coronaRedisRepository.deleteAll();
         clearAllNewCaseCountCache();
-
-        ICoronaInfo info = govCoronaOpenApi.getInfo();
 
         for (int i = 0; i < info.getItemList().size(); i++) {
             ICoronaItem item = info.getItemList().get(i);
@@ -107,7 +112,7 @@ public class CoronaService {
         log.info("fetchAndStoreCorona::코로나 데이터를 성공적으로 갱신하였습니다.");
     }
 
-    public void convertInfoToScore(CoronaInfo coronaInfo, ScoreResultResponseDto scoreResultResponseDto) {
+    public void convertInfoToScore(ScoreResultResponseDto scoreResultResponseDto) {
         final int CORONA_LEVEL15 = 300;
         final int CORONA_LEVEL2 = 400;
         final int CORONA_LEVEL25 = 800;
